tagstruct: copy: Recognize _APPENDED types
If we're copying from a PA_TAGSTRUCT_APPENDED tagstruct, we don't
need any extra copy for the data: all data is originally appended
to the source and thus already-copied in the destination.

Do not do the same for fixed tagstructs: they don't own the data
and we don't want to create dangling references behind original
owner's back.

Suggested-by: Peter Meerwald <pmeerw@pmeerw.net>
Signed-off-by: Ahmed S. Darwish <darwish.07@gmail.com>