pulsecore: srbchannel: Extract ringbuffer code into its own file
To solve a 10.0 release blocker, pacat is to use this ringbuffer
implementation for its own playback mode.

Matching the rest of global pulse code, the ring initialization
APIs should be more opaque and sizes should be in size_t rather
than integers. Nonetheless, they're left as is out of care not
to modify the critical srbchannel code this late in the process.

While at it, further document the ringbuffer.h prototypes.

Signed-off-by: Ahmed S. Darwish <darwish.07@gmail.com>